+++ tests/test_main.py
@@ -1,4 +1,28 @@
-"""Test the :mod:`psyplot.__main__` module"""
+"""Test the :mod:`psyplot.__main__` module."""
+
+# Disclaimer
+# ----------
+#
+# Copyright (C) 2021 Helmholtz-Zentrum Hereon
+# Copyright (C) 2020-2021 Helmholtz-Zentrum Geesthacht
+# Copyright (C) 2016-2021 University of Lausanne
+#
+# This file is part of psyplot and is released under the GNU LGPL-3.O license.
+# See COPYING and COPYING.LESSER in the root of the repository for full
+# licensing details.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License version 3.0 as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU LGPL-3.0 license for more details.
+#
+# You should have received a copy of the GNU LGPL-3.0 license
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import sys
 import os
 import os.path as osp
@@ -143,6 +167,11 @@ class TestCommandLine(unittest.TestCase):
         d = yaml.load(proc.stdout.read(), yaml.Loader)
         d.pop('psyplot_gui', None)
         ref.pop('psyplot_gui', None)
+        # make sure the version does not end with .dirty
+        d["psyplot"]["version"] = d["psyplot"]["version"].replace(".dirty", "")
+        ref["psyplot"]["version"] = ref["psyplot"]["version"].replace(
+            ".dirty", ""
+        )
         self.assertEqual(d, ref)
 
     def test_list_plugins(self):